mind/requirements2.txt
about-time==4.2.1
alive-progress==3.1.5
astunparse==1.6.3
atomicwrites==1.4.0
autograd==1.7.0
cma==3.2.2
grapheme==0.6.0
inflection==0.5.1
Mako==1.3.10
Markdown==3.7
matplotlib==3.3.4
mkl-service==2.4.0
nose==1.3.7
numpy==1.24.4
pandas==1.2.1
pdoc==14.7.0
pdoc3==0.9.2
ply==3.11
pyls-spyder==0.4.0
pymoo==0.6.1.2
Pyomo==5.7.3
PyQt5==5.15.10
PyQtWebEngine==5.15.6
PyUtilib==6.0.0
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
sklearn==0.0
webencodings==0.5.1

